MNIST/requirements.txt
torch 
torchvision 
onnx 
onnxruntime 
pillow 
numpy 
matplotlib 
jupyter